% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_res.R
\name{varyres}
\alias{varyres}
\title{A variable-resolution heat map generator}
\usage{
varyres(dataset, cutoff)
}
\arguments{
\item{dataset}{data frame with spatial data: x-coordinates (px), y coordinates (pz), and Bernoulli responses at those locations (hit)}

\item{cutoff}{Box subdivisions cease when a box sample size drops below the cutoff}
}
\description{
This function creates variable resolution heat maps accroding to a stopping rule
}
\examples{
data(dataset)
data <- varyres(dataset, 200)
mapit(data[[4]])
}
